(test) Refactor tests to use naming conventions and BDD-like syntax
