+++ e2e/commands/encounter-operations.ts
