+++ e2e/commands/index.ts
@@ -1,5 +1,5 @@
-export * from './patientOperations';
-export * from './encounterOperations';
-export * from './visitOperations';
-export * from './providerOperations';
-export * from './cohortOperations';
+export * from './patient-operations';
+export * from './encounter-operations';
+export * from './visit-operations';
+export * from './provider-operations';
+export * from './cohort-operations';
