+++ e2e/commands/patient-operations.ts
