+++ e2e/commands/provider-operations.ts
