+++ e2e/commands/visit-operations.ts
