+++ e2e/pages/registration-and-edit-page.ts
